dashboard_requirements.txt
flask-cors==4.0.0
flask-socketio==5.3.6
ckanapi==4.7
pandas==2.1.1
python-dotenv==1.0.0
schedule==1.2.0
slack-sdk==3.21.3